feat: Set "Blend" as default editor tab
This commit changes the default selected tab in the `EditorScreen` from "Image 1" to "Blend".

In the `EditorTabs` composable, the initial state for `tabIndex` has been updated from `0` to `2`. This makes the "Blend" tab the active tab when the editor is first displayed, prioritizing the view that shows the combined result of the two images.